pyproject.toml
[build-system]
requires = ["setuptools>=68", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "ezqmmm"
version = "2.0.1"
description = "Easy QM/MM input file generator from CHARMM/NAMD MD trajectories"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "ezQMMM Contributors" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Chemistry",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "MDAnalysis>=2.4",
    "numpy>=1.22",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4",
]

[project.scripts]
ezqmmm = "ezqmmm.cli:main"

[tool.setuptools.packages.find]
include = ["ezqmmm*"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "W",    # pycodestyle warnings
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
]
ignore = [
    "E501",   # line too long — handled by formatter
    "E741",   # ambiguous variable name (l, O, I) — common in science code
    "B006",   # mutable default argument — false positives on numpy arrays
    "SIM108", # ternary operator — readability matters more here
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["E501", "B011"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: marks tests that load trajectories (deselect with '-m \"not slow\"')",
]
